+++ packages/be-ts-db-lib/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [4.2.4] (2026-02-17)
+
+### 🏗️ Changed
+
+- Updated dependencies: now requires `@aneuhold/be-ts-lib@^3.0.14` and `@aneuhold/core-ts-db-lib@^4.1.3`.
+
+### 🔥 Removed
+
+- Removed outdated `WORKOUT_IMPLEMENTATION_PLAN.md` documentation file.
+
 ## 🔖 [4.2.3] (2026-02-15)
 
 ### 🏗️ Changed
@@ -233,6 +243,7 @@ Updated dependencies: now requires `@aneuhold/core-ts-db-lib@^3.0.0`, `@aneuhold
 
 <!-- Link References -->
 
+[4.2.4]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.3...be-ts-db-lib-v4.2.4
 [4.2.3]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.2...be-ts-db-lib-v4.2.3
 [4.2.2]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.1...be-ts-db-lib-v4.2.2
 [4.2.1]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.0...be-ts-db-lib-v4.2.1